+++ src/utils/umath.hpp
@@ -16,26 +16,26 @@ T urand_t(T min, T max) {
 }
 
 template<typename T>
-void rnd_vector(const ulong size, std::vector<T> &v, const T min, const T max) {
-    for (int i = 0; i < size; ++i) {
+void rnd_vector(const uint64_t size, std::vector<T> &v, const T min, const T max) {
+    for (size_t i = 0; i < size; ++i) {
         v.emplace_back(urand_t(min, max));
     }
 }
 
 template<typename T>
-void zeros_vector(const ulong size, std::vector<T> &v) {
-    for (int i = 0; i < size; ++i) {
+void zeros_vector(const uint64_t size, std::vector<T> &v) {
+    for (size_t i = 0; i < size; ++i) {
         v.emplace_back(0);
     }
 }
 
 template<typename T>
-void ddm(const ulong size, std::vector<std::vector<T> > &matrix,
+void ddm(const uint64_t size, std::vector<std::vector<T> > &matrix,
                                        const T min, const T max) {
-    for (ulong i = 0; i < size; ++i) {
+    for (size_t i = 0; i < size; ++i) {
         T sum = T(0);
         std::vector<T> row_vec;
-        for (ulong j = 0; j < size; ++j){
+        for (size_t j = 0; j < size; ++j){
             T value = urand_t(min, max);
             sum += std::abs(value);
             row_vec.emplace_back(value);
@@ -50,10 +50,10 @@ template<typename T>
 bool check_ddm(std::vector<std::vector<T> > &matrix){
 	int check_count = 0;
 	// For each row ..
-	for (int i = 0; i < matrix.size(); i++) {
+	for (size_t i = 0; i < matrix.size(); i++) {
 		float row_sum = 0;
 		// Summing the other row elements .. 
-		for (int j = 0; j < matrix.size(); j++) {
+		for (size_t j = 0; j < matrix.size(); j++) {
 			if (j != i) row_sum += abs(matrix[i][j]);
 		}
